chore/msp-example: refactor to align with service structure best practices (#62954)
In monorepo:

```
cmd/my-service/main.go
-> cmd/my-service/service
-> cmd/my-service/internal/...
```

Outside monorepo a similar unnested structure aligns with this as well:

```
cmd/my-service <- command
service/...    <- entrypoint
internal/...   <- internal implementation
```


## Test plan

Basic example builds and runs: `sg run msp-example`
