--- cmd/msp-example/internal/example/BUILD.bazel
@@ -1,31 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "example",
-    srcs = [
-        "bigquery.go",
-        "diagnostics.go",
-        "example.go",
-        "metrics.go",
-        "postgresql.go",
-        "redis.go",
-    ],
-    importpath = "github.com/sourcegraph/sourcegraph/cmd/msp-example/internal/example",
-    tags = [TAG_INFRA_CORESERVICES],
-    visibility = [
-        "//cmd/msp-example:__pkg__",
-        "//lib/managedservicesplatform/example:__subpackages__",
-    ],
-    deps = [
-        "//internal/version",
-        "//lib/background",
-        "//lib/errors",
-        "//lib/managedservicesplatform/runtime",
-        "@com_github_go_redis_redis_v8//:redis",
-        "@com_github_sourcegraph_log//:log",
-        "@com_google_cloud_go_bigquery//:bigquery",
-        "@io_opentelemetry_go_contrib_instrumentation_net_http_otelhttp//:otelhttp",
-        "@io_opentelemetry_go_otel//:otel",
-        "@io_opentelemetry_go_otel_metric//:metric",
-    ],
-)
