+++ cmd/msp-example/internal/redis/BUILD.bazel
@@ -1,0 +1,13 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "redis",
+    srcs = ["redis.go"],
+    importpath = "github.com/sourcegraph/sourcegraph/cmd/msp-example/internal/redis",
+    visibility = ["//cmd/msp-example:__subpackages__"],
+    deps = [
+        "//lib/errors",
+        "//lib/managedservicesplatform/runtime",
+        "@com_github_go_redis_redis_v8//:redis",
+    ],
+)
